ISSUE #4731 - fix: circular dependency
